+++ vulnerabilities/tests/pipelines/test_enhance_with_kev.py
@@ -1,3 +1,12 @@
+#
+# Copyright (c) nexB Inc. and others. All rights reserved.
+# VulnerableCode is a trademark of nexB Inc.
+# SPDX-License-Identifier: Apache-2.0
+# See http://www.apache.org/licenses/LICENSE-2.0 for the license text.
+# See https://github.com/aboutcode-org/vulnerablecode for support or download.
+# See https://aboutcode.org for more information about nexB OSS projects.
+#
+
 import os
 from unittest import mock
 from unittest.mock import Mock
@@ -7,29 +16,32 @@ import pytest
 from vulnerabilities.models import Alias
 from vulnerabilities.models import Exploit
 from vulnerabilities.models import Vulnerability
-from vulnerabilities.pipelines.enhance_with_metasploit import MetasploitImproverPipeline
+from vulnerabilities.pipelines.enhance_with_kev import VulnerabilityKevPipeline
 from vulnerabilities.utils import load_json
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-TEST_DATA = os.path.join(BASE_DIR, "../test_data", "metasploit_improver/modules_metadata_base.json")
+TEST_DATA = os.path.join(BASE_DIR, "../test_data", "kev_data.json")
 
 
 @pytest.mark.django_db
 @mock.patch("requests.get")
-def test_metasploit_improver(mock_get):
+def test_kev_improver(mock_get):
     mock_response = Mock(status_code=200)
     mock_response.json.return_value = load_json(TEST_DATA)
     mock_get.return_value = mock_response
 
-    improver = MetasploitImproverPipeline()
+    improver = VulnerabilityKevPipeline()
 
     # Run the improver when there is no matching aliases
     improver.execute()
+
     assert Exploit.objects.count() == 0
 
     v1 = Vulnerability.objects.create(vulnerability_id="VCIO-123-2002")
-    Alias.objects.create(alias="CVE-2007-4387", vulnerability=v1)
+    v1.save()
 
-    # Run metasploit Improver again when there are matching aliases.
+    Alias.objects.create(alias="CVE-2021-38647", vulnerability=v1)
+
+    # Run Kev Improver again when there are matching aliases.
     improver.execute()
     assert Exploit.objects.count() == 1
